Add files via upload
As a novice, I initially struggled to grasp the concept of using 'zip.' Recognizing that others might encounter similar difficulties, I'd like to share an alternative method that simplifies the process of calculating the output. This approach proved helpful in enhancing my understanding, and I hope it proves helpful to fellow learners.